Fix size_t comparision from gsl_span.  Fix mlas merge issue.  Revert one of the std::ref_wrap usage.

+++ onnxruntime/core/framework/op_node_proto_helper.cc
@@ -80,7 +80,7 @@ inline bool HasTyped<GraphProto>(const AttributeProto* attr) {
     if (!attr) {                                                                   \
       return Status(ONNXRUNTIME, FAIL, "No attribute with this name is defined."); \
     }                                                                              \
-    ORT_ENFORCE(values.size() == attr->list##_size());                             \
+    ORT_ENFORCE(values.size() == static_cast<size_t>(attr->list##_size()));        \
     for (int i = 0; i < attr->list##_size(); ++i) {                                \
       values[i] = static_cast<T>(attr->list(i));                                   \
     }                                                                              \
